Basic Spring Boot Framework and Courses Class
I updated the Spring Boot file with the most recent HTML and JS files. I created test classes for Product and Student. We do not need these classes, but they have the blueprint of creating classes going forward. I created the Course class and the corresponding supporting classes, CourseController, CourseRepository, and CourseService to handle the web, server, and business layers of the Course objects/beans.

The Course class was using information from a demo server I created, so the application.properties file will need to be updated with the current information. The Courses class will also need to be updated with the current table contents from the actual server.