Added Filtering for Student Courses taken and Student Repository
For Optimizing Spring Boot APIs, I added filtering to the StudentRepository so that it only shows information for the student logged in. For Cross Referencing Classes Taken to Pre-reqs needed for each class, I added a ManyToMany set of courses taken so that the student could retain which courses they have taken to be cross referenced to the major_requirements.